+++ client/aux_test.go
@@ -1,4 +1,4 @@
-package tx_test
+package client_test
 
 import (
 	"context"
